implemented seperate api auth middleware
