+++ middlewares/withWorkspaceAuth.ts
@@ -1,10 +1,10 @@
 /* eslint-disable indent */
 import { User, Workspace } from '@prisma/client';
 import { NextRequest } from 'next/server';
+import { withApiAuth } from './withApiAuth';
 import { isEmpty } from '@/helpers/common';
 import { getSessionDetails } from '@/services/serverSide/auth/authentication';
 import { prisma } from '@/prisma/prisma';
-import { getApiDetails } from '@/services/serverSide/apiKey';
 
 type AuthorizedRequest = NextRequest & { user: User; workspace: Workspace };
 // eslint-disable-next-line no-unused-vars
@@ -22,29 +22,7 @@ const withWorkspaceAuth =
     const authorizedRequest = req as AuthorizedRequest;
 
     if (authorizationHeader) {
-      if (!authorizationHeader.startsWith('Bearer ')) {
-        return Response.json(
-          {
-            error:
-              "Misconfigured authorization header. Did you forget to add 'Bearer '?",
-          },
-          { status: 400 },
-        );
-      }
-
-      const apiKey = authorizationHeader.slice(7);
-
-      const apiKeyDetails = await getApiDetails(apiKey);
-
-      if (!apiKeyDetails) {
-        return Response.json(
-          { error: 'Unauthorized: Invalid API key!' },
-          { status: 401 },
-        );
-      }
-
-      authorizedRequest.user = apiKeyDetails.user;
-      authorizedRequest.workspace = apiKeyDetails.workspace;
+      return withApiAuth(handler);
     } else {
       const session = await getSessionDetails();
